+++ packages/gestalt/src/IconButton.js
@@ -1,6 +1,7 @@
 // @flow
 import * as React from 'react';
 import PropTypes from 'prop-types';
+import classnames from 'classnames';
 import styles from './IconButton.css';
 import icons from './icons/index.js';
 import Pog from './Pog.js';
@@ -15,8 +16,10 @@ type Props = {|
     | 'gray'
     | 'lightGray'
     | 'white',
+  dangerouslySetSvgPath?: { __path: string },
+  disabled?: boolean,
   iconColor?: 'gray' | 'darkGray' | 'red' | 'blue' | 'white',
-  icon: $Keys<typeof icons>,
+  icon?: $Keys<typeof icons>,
   onClick?: ({ event: SyntheticMouseEvent<> }) => void,
   size?: 'xs' | 'sm' | 'md' | 'lg' | 'xl',
 |};
@@ -39,7 +42,11 @@ export default class IconButton extends React.Component<Props, State> {
       'lightGray',
       'white',
     ]),
-    icon: PropTypes.oneOf(Object.keys(icons)).isRequired,
+    dangerouslySetSvgPath: PropTypes.shape({
+      __path: PropTypes.string,
+    }),
+    disabled: PropTypes.bool,
+    icon: PropTypes.oneOf(Object.keys(icons)),
     iconColor: PropTypes.oneOf(['gray', 'darkGray', 'red', 'blue', 'white']),
     onClick: PropTypes.func,
     size: PropTypes.oneOf(['xs', 'sm', 'md', 'lg', 'xl']),
@@ -69,6 +76,8 @@ export default class IconButton extends React.Component<Props, State> {
       accessibilityHaspopup,
       accessibilityLabel,
       bgColor,
+      dangerouslySetSvgPath,
+      disabled,
       iconColor,
       icon,
       size,
@@ -82,7 +91,11 @@ export default class IconButton extends React.Component<Props, State> {
         aria-expanded={accessibilityExpanded}
         aria-haspopup={accessibilityHaspopup}
         aria-label={accessibilityLabel}
-        className={styles.button}
+        className={classnames(
+          styles.button,
+          disabled ? styles.disabled : styles.enabled
+        )}
+        disabled={disabled}
         onBlur={this.handleBlur}
         onClick={event => onClick && onClick({ event })}
         onFocus={this.handleFocus}
@@ -93,10 +106,11 @@ export default class IconButton extends React.Component<Props, State> {
         type="button"
       >
         <Pog
-          active={active}
+          active={!disabled && active}
           bgColor={bgColor}
-          focused={focused}
-          hovered={hovered}
+          dangerouslySetSvgPath={dangerouslySetSvgPath}
+          focused={!disabled && focused}
+          hovered={!disabled && hovered}
           iconColor={iconColor}
           icon={icon}
           size={size}