+++ packages/gestalt/src/Spinner.js
@@ -1,6 +1,7 @@
 // @flow
 import * as React from 'react';
 import PropTypes from 'prop-types';
+import classnames from 'classnames';
 import Box from './Box.js';
 import Icon from './Icon.js';
 import styles from './Spinner.css';
@@ -9,13 +10,18 @@ const SIZE = 40;
 
 type Props = {|
   accessibilityLabel: string,
+  delay?: boolean,
   show: boolean,
 |};
 
-export default function Spinner({ accessibilityLabel, show }: Props) {
+export default function Spinner({
+  accessibilityLabel,
+  delay = true,
+  show,
+}: Props) {
   return show ? (
     <Box display="flex" justifyContent="around" overflow="hidden">
-      <div className={styles.icon}>
+      <div className={classnames(styles.icon, { [styles.delay]: delay })}>
         <Icon
           icon="knoop"
           accessibilityLabel={accessibilityLabel}
@@ -31,4 +37,5 @@ export default function Spinner({ accessibilityLabel, show }: Props) {
 Spinner.propTypes = {
   show: PropTypes.bool.isRequired,
   accessibilityLabel: PropTypes.string.isRequired,
+  delay: PropTypes.bool,
 };