+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Minor
 
+- Eslint: Upgrade to `v6.0.0` and bump all related packages/plugins (#544)
+
 ### Patch
 
 </details>