+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Minor
 
+- Enzyme: Upgrade to the latest `v3.10.0` version and pull in Flow library changes (#543)
+
 ### Patch
 
 </details>